Move `syntax` into class `syntactic_environment` from `syntax.hpp`
