--- include/meevax/kernel/syntax.hpp
@@ -1,47 +1,0 @@
-/*
-   Copyright 2018-2022 Tatsuya Yamasaki.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-#ifndef INCLUDED_MEEVAX_KERNEL_SYNTAX_HPP
-#define INCLUDED_MEEVAX_KERNEL_SYNTAX_HPP
-
-#include <meevax/kernel/pair.hpp>
-
-namespace meevax
-{
-inline namespace kernel
-{
-  class environment; // environment.hpp
-
-  struct syntax
-  {
-    using function_type = std::function<auto (environment &,
-                                              object const&,
-                                              object const&,
-                                              object const&,
-                                              object const&) -> object>;
-
-    std::string const name;
-
-    function_type const compile;
-
-    explicit syntax(std::string const&, function_type const&);
-  };
-
-  auto operator <<(std::ostream &, syntax const&) -> std::ostream &;
-} // namespace kernel
-} // namespace meevax
-
-#endif // INCLUDED_MEEVAX_KERNEL_SYNTAX_HPP
